requirements.txt
# List your project dependencies here
pdfminer.six
python-docx
Pillow
